DNSSEC refactoring
* Better encapsulation and responsibility of entities
* Move tests to appropriate files
* Better naming
